Add display functionality on calculator
I choose to add a unique event listener to each
button to display the number being clicked
even though the functionality is the same
for all digits except 0. In the future, I think
there is an opportunity to refactor this into one
function that takes the button info as arguments,
but currently that makes my head hurt haha.