Update calculator styling
Updating styling now as it will make it easier to
write the logic to round numbers so they don't
extend past the display. Also, choose to use
flexbox as I do not believe we should be using
forms in this challenge; however, I would have
likely used forms as I think it would have been
easier to have everything in their own rows for
submitting.