Add logic for rounding calculator display values
Choose to use a blank else if statement if the
values were >= 17 to stop overflow for extra digits
being entered. For the operators, a mix of string
and number methods were used to control the rounding
of large decimals. There may be an easier or better way,
but this is what I found to work.