Fix logic to clear display with a new digit press
Choose to add a new boolean variable to determine
if the result was after the equals button or an
operator. This made it easy to specify that a new
key press after a result from a equals button would
clear the display and display the new value.
Essentially, this allows the operator buttons to
display a result and capture that result for the next
operation, whereas the equals button allows only a
result that will be cleared with the next digit.