Add keyboard support
Decided to just copy the current logic from the buttons
and moved the same logic inside a keydown event handler.

In the future, there is definitely an opportunity to refractor
to make the logic into functions that can be passed to both the
button and keydown handlers. Additionally, I am sure there is a
way to dynamically pass in the right values and numbers for the
buttons and keydowns so likely can clean this up a good deal.